preventing overflow for high numbers of particles and number of interactions. Thanks to Yun Zhang for finding!

+++ miluph.cu
@@ -96,7 +96,7 @@ int maxNumberOfParticles;
 
 int memorySizeForTree;
 int memorySizeForParticles;
-int memorySizeForInteractions;
+size_t memorySizeForInteractions;
 int memorySizeForChildren;
 int memorySizeForStress;
 #if FRAGMENTATION
